requirements.txt
PySide6>=6.6.0
requests>=2.31.0
tomli-w>=1.0.0
tomlkit>=0.12.0
websocket-client>=1.6.0
py7zr>=0.20.0
rarfile>=4.0
pyinstaller>=6.0.0
